+++ packages/sdks-tests/src/e2e-tests/ab-test.spec.ts
@@ -1,6 +1,7 @@
 import type { Browser } from '@playwright/test';
 import { expect } from '@playwright/test';
-import { excludeTestFor, test } from '../helpers/index.js';
+import { test } from '../helpers/index.js';
+
 const SELECTOR = 'div[builder-content-id]';
 
 const createContextWithCookies = async ({
@@ -78,19 +79,13 @@ test.describe('A/B tests', () => {
 
     const TRIES = 10;
 
-    // Manually run tests 10 times to ensure we don't have any flakiness.
     for (let i = 1; i <= TRIES; i++) {
       test(`#${i}/${TRIES}: Render default w/ SSR`, async ({
         page: _page,
         baseURL,
         packageName,
         browser,
-        sdk,
       }) => {
-        test.skip(
-          excludeTestFor({ angular: true }, sdk),
-          'A/B tests with SSR are not supported in Angular'
-        );
         const { page } = await initializeAbTest(
           {
             page: _page,
@@ -104,7 +99,28 @@ test.describe('A/B tests', () => {
           }
         );
 
-        await page.goto('/ab-test');
+        let trackCalls = 0;
+
+        await page.route('**/track**', async route => {
+          const url = route.request().url();
+          if (url.includes('cdn.builder.io/api/v1/track')) {
+            trackCalls += 1;
+            const payload = route.request().postDataJSON();
+            if (payload.events[0].data.variationId) {
+              throw new Error(
+                'Unexpected variationId in track request payload for default variant'
+              );
+            }
+            if (payload.events[0].data.contentId !== CONTENT_ID) {
+              throw new Error('ContentId does not match expected default contentId');
+            }
+          }
+          await route.continue();
+        });
+
+        await page.goto('/ab-test', { waitUntil: 'networkidle' });
+
+        expect(trackCalls).toBe(1);
 
         await expect(page.getByText(TEXTS.DEFAULT_CONTENT).locator('visible=true')).toBeVisible();
         await expect(page.locator(SELECTOR, { hasText: TEXTS.VARIANT_1 })).toBeHidden();
@@ -116,12 +132,7 @@ test.describe('A/B tests', () => {
         baseURL,
         packageName,
         browser,
-        sdk,
       }) => {
-        test.skip(
-          excludeTestFor({ angular: true }, sdk),
-          'A/B tests with SSR are not supported in Angular'
-        );
         const { page } = await initializeAbTest(
           {
             page: _page,
@@ -134,8 +145,22 @@ test.describe('A/B tests', () => {
             cookieValue: VARIANT_ID,
           }
         );
+        let trackCalls = 0;
+        await page.route('**/track**', async route => {
+          const url = route.request().url();
+          if (url.includes('cdn.builder.io/api/v1/track')) {
+            trackCalls += 1;
+            const payload = route.request().postDataJSON();
+            if (!payload.events || payload.events[0].data.variationId !== VARIANT_ID) {
+              throw new Error('Missing variationId in track request payload');
+            }
+          }
+          await route.continue();
+        });
 
-        await page.goto('/ab-test');
+        await page.goto('/ab-test', { waitUntil: 'networkidle' });
+
+        expect(trackCalls).toBe(1);
 
         await expect(page.getByText(TEXTS.VARIANT_1).locator('visible=true')).toBeVisible();
         await expect(page.locator(SELECTOR, { hasText: TEXTS.DEFAULT_CONTENT })).toBeHidden();
@@ -164,12 +189,7 @@ test.describe('A/B tests', () => {
         baseURL,
         packageName,
         browser,
-        sdk,
       }) => {
-        test.skip(
-          excludeTestFor({ angular: true }, sdk),
-          'Nested symbols are not supported in Angular'
-        );
         const { page } = await initializeAbTest(
           {
             page: _page,
@@ -198,12 +218,7 @@ test.describe('A/B tests', () => {
         baseURL,
         packageName,
         browser,
-        sdk,
       }) => {
-        test.skip(
-          excludeTestFor({ angular: true }, sdk),
-          'Nested symbols are not supported in Angular'
-        );
         const { page } = await initializeAbTest(
           {
             page: _page,