+++ packages/sdks-tests/src/specs/video-lazy-load.ts
@@ -1,0 +1,68 @@
+export const VIDEO_LAZY_LOAD = {
+  data: {
+    title: 'react-page-1',
+    themeId: false,
+    blocks: [
+      {
+        '@type': '@builder.io/sdk:Element',
+        '@version': 2,
+        layerName: 'Texttttt.....',
+        id: 'builder-7c03ea33c7f149c89436e16cc9e233db',
+        component: {
+          name: 'Text',
+          options: { text: 'afsdfsdfEnter some text...' },
+        },
+        responsiveStyles: {
+          large: {
+            display: 'flex',
+            flexDirection: 'column',
+            position: 'relative',
+            flexShrink: '0',
+            boxSizing: 'border-box',
+            lineHeight: 'normal',
+            height: 'auto',
+            marginTop: '2000px',
+          },
+        },
+      },
+      {
+        '@type': '@builder.io/sdk:Element',
+        '@version': 2,
+        id: 'builder-863a7f6160e943f2ae54a3b7ac6bf0ef',
+        component: {
+          name: 'Video',
+          options: {
+            video:
+              'https://cdn.builder.io/o/assets%2Fad30f9a246614faaa6a03374f83554c9%2Fb637ba8975044a8fa751c9386c281b77%2Fcompressed?apiKey=ad30f9a246614faaa6a03374f83554c9&token=b637ba8975044a8fa751c9386c281b77&alt=media&optimized=true',
+            autoPlay: true,
+            controls: false,
+            muted: true,
+            loop: true,
+            playsInline: true,
+            fit: 'cover',
+            preload: 'metadata',
+            fitContent: true,
+            position: 'center',
+            aspectRatio: 0.7004048582995948,
+            lazyLoad: true,
+            posterImage:
+              'https://cdn.builder.io/api/v1/image/assets%2F101392854f9a4c9cb60fbd1d98aa3fc3%2Fb8ce092dd2124b2c854f2c10b51f7597',
+          },
+        },
+        responsiveStyles: {
+          large: {
+            display: 'flex',
+            flexDirection: 'column',
+            position: 'relative',
+            flexShrink: '0',
+            boxSizing: 'border-box',
+            marginTop: '20px',
+            minHeight: '20px',
+            minWidth: '20px',
+            width: '100%',
+          },
+        },
+      },
+    ],
+  },
+};