+++ packages/sdks/snippets/react/src/main.tsx
@@ -6,6 +6,7 @@ import AdvancedChildRoute from './routes/custom-components/advanced-child.tsx';
 import CustomChildRoute from './routes/custom-components/custom-child.tsx';
 import EditableRegionRoute from './routes/custom-components/editable-region.tsx';
 import IntegratingPages from './routes/IntegratingPages.tsx';
+import LivePreviewBlogData from './routes/LivePreviewBlogData.js';
 
 const router = createBrowserRouter([
   {
@@ -23,6 +24,10 @@ const router = createBrowserRouter([
   {
     path: '/advanced-child',
     element: <AdvancedChildRoute />,
+  },
+  {
+    path: '/live-preview',
+    element: <LivePreviewBlogData />,
   },
   {
     path: '/*',