+++ packages/sdks/src/blocks/text/text.lite.tsx
@@ -1,4 +1,11 @@
+import { useMetadata } from '@builder.io/mitosis';
 import type { TextProps } from './text.types.js';
+
+useMetadata({
+  angular: {
+    changeDetection: 'OnPush',
+  },
+});
 
 export default function Text(props: TextProps) {
   return (