+++ packages/sdks/output/angular/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @builder.io/sdk-angular
 
+## 0.18.4
+
+### Patch Changes
+
+- 887c6e0: Fix: visual editing Custom Code block reflects code updates in real time
+- 887c6e0: Fix: `nativeElement` not found issue in Custom Code and Embed Blocks
+
+## 0.18.3
+
+### Patch Changes
+
+- 527c115: Feat: A/B tests support
+
+## 0.18.2
+
+### Patch Changes
+
+- 1396fb4: Fix: duplicate `/track` call validation handles default and variant scenarios correctly.
+
+## 0.18.1
+
+### Patch Changes
+
+- 80247eb: Fix: Updating input values while in the Content Editor not triggering changes in the iframe.
+
+## 0.18.0
+
+### Minor Changes
+
+- 5ed08fc: - BREAKING CHANGE 🧨 : updated `subscribeToEditor` arguments: - arguments are now passed as a named argument object - `apiKey` is now a required field
+
+Example:
+
+- from:
+
+```ts
+    subscribeToEditor('page', () => { ... }, options: {trustedHosts:['...']})
+```
+
+- to:
+
+```ts
+    subscribeToEditor({
+        apiKey: '...',
+        model: '...',
+        trustedHosts: ['...'],
+        callback: () => { ... }
+    })
+```
+
+- 10a5754: BREAKING CHANGE 🧨: `model` and `content` are now required props for `<Content>`.
+
+## 0.17.9
+
+### Patch Changes
+
+- 58ee59e: Fix: added lazy loading to video element
+
 ## 0.17.8
 
 ### Patch Changes